requirements.txt
cachetools==4.2.2
DateTime==4.3
google-auth==1.34.0
google-auth-oauthlib==0.4.4
gspread==3.7.0
oauthlib==3.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2021.1
requests-oauthlib==1.3.0
rsa==4.7.2
zope.interface==5.4.0
